pyproject.toml
[project]
name = "efflux-desktop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "anthropic>=0.49.0",
    "autogen-agentchat>=0.5.7",
    "autogen-core>=0.5.7",
    "boto3>=1.38.43",
    "cachetools>=5.5.2",
    "colorlog>=6.9.0",
    "docker>=7.1.0",
    "fastapi>=0.115.12",
    "google-genai>=1.22.0",
    "ijson>=3.4.0",
    "injector>=0.22.0",
    "jsonlines>=4.0.0",
    "loguru>=0.7.3",
    "markitdown>=0.1.1",
    "mcp>=1.6.0",
    "openai>=1.72.0",
    "pdfminer-six>=20250327",
    "pillow>=11.2.1",
    "playwright>=1.52.0",
    "pyinstaller>=6.14.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "socksio>=1.0.0",
    "tiktoken>=0.9.0",
    "tldextract>=5.3.0",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]
